Introduce actions functionality and update Invoices UI
Implemented functions for invoice actions such as create, update, and delete within actions.ts. Introduced the capability for form state management in the invoices "edit-form". Updated the operations on the Invoices buttons as well as improved the UI of the Invoices editing page. Alongside, dependencies in package.json have been updated.